# Speedup HashTrie iteration