Speedup HashTrie Iterator by:
* avoid pushing single bucket leafs into the stack
* avoid recursion
* avoid iterator peeking

Before:
rpds hash trie map sync iterate
                        time:   [301.18 us 304.05 us 307.06 us]

After (this PR, non inlined next)
rpds hash trie map sync iterate
                        time:   [120.29 us 120.96 us 121.72 us]

After (when next is inlined, not in this PR)
rpds hash trie map sync iterate
                        time:   [103.74 us 104.39 us 105.03 us]