@orium 2021-09-27
Thanks for the contribution :)

I'll make a release soon.